boardd: prep for SPI + factor out USB (#26356)
* merge origin/spi-panda

* just prep

* boardd: factor out USB comms

* fix those

* add to release files

* little more
