+++ release/files_common
@@ -90,6 +90,8 @@ selfdrive/boardd/boardd_api_impl.pyx
 selfdrive/boardd/can_list_to_can_capnp.cc
 selfdrive/boardd/panda.cc
 selfdrive/boardd/panda.h
+selfdrive/boardd/panda_comms.h
+selfdrive/boardd/panda_comms.cc
 selfdrive/boardd/set_time.py
 selfdrive/boardd/pandad.py
