+++ selfdrive/boardd/panda.h
@@ -1,26 +1,26 @@
 #pragma once
 
-#include <atomic>
 #include <cstdint>
 #include <ctime>
 #include <functional>
 #include <list>
-#include <mutex>
+#include <memory>
 #include <optional>
 #include <vector>
-
-#include <libusb-1.0/libusb.h>
 
 #include "cereal/gen/cpp/car.capnp.h"
 #include "cereal/gen/cpp/log.capnp.h"
 #include "panda/board/health.h"
+#include "selfdrive/boardd/panda_comms.h"
 
-#define TIMEOUT 0
+
 #define PANDA_CAN_CNT 3
 #define PANDA_BUS_CNT 4
-#define RECV_SIZE (0x4000U)
+
 #define USB_TX_SOFT_LIMIT   (0x100U)
 #define USBPACKET_MAX_SIZE  (0x40)
+
+#define RECV_SIZE (0x4000U)
 #define CANPACKET_HEAD_SIZE 5U
 #define CANPACKET_MAX_SIZE  72U
 #define CANPACKET_REJECTED  (0xC0U)
@@ -37,40 +37,31 @@ struct __attribute__((packed)) can_header {
 };
 
 struct can_frame {
-	long address;
-	std::string dat;
-	long busTime;
-	long src;
+  long address;
+  std::string dat;
+  long busTime;
+  long src;
 };
 
+
 class Panda {
- private:
-  libusb_context *ctx = NULL;
-  libusb_device_handle *dev_handle = NULL;
-  std::mutex usb_lock;
+private:
+  std::unique_ptr<PandaCommsHandle> handle;
   std::vector<uint8_t> recv_buf;
-  void handle_usb_issue(int err, const char func[]);
-  void cleanup();
 
- public:
+public:
   Panda(std::string serial="", uint32_t bus_offset=0);
-  ~Panda();
 
-  std::string usb_serial;
-  std::atomic<bool> connected = true;
-  std::atomic<bool> comms_healthy = true;
+  std::string hw_serial;
   cereal::PandaState::PandaType hw_type = cereal::PandaState::PandaType::UNKNOWN;
   bool has_rtc = false;
   const uint32_t bus_offset;
 
+  bool connected();
+  bool comms_healthy();
+
   // Static functions
   static std::vector<std::string> list();
-
-  // HW communication
-  int usb_write(uint8_t bRequest, uint16_t wValue, uint16_t wIndex, unsigned int timeout=TIMEOUT);
-  int usb_read(uint8_t bRequest, uint16_t wValue, uint16_t wIndex, unsigned char *data, uint16_t wLength, unsigned int timeout=TIMEOUT);
-  int usb_bulk_write(unsigned char endpoint, unsigned char* data, int length, unsigned int timeout=TIMEOUT);
-  int usb_bulk_read(unsigned char endpoint, unsigned char* data, int length, unsigned int timeout=TIMEOUT);
 
   // Panda functionality
   cereal::PandaState::PandaType get_hw_type();
